Refactor NetworkInstance methods into NetworkInstanceMixin for better code organization
